fix(assembler): Fix wrong address for labels
The assembler was generating wrong address for label with NOP
instruction. This was because NOP was not occupying any space in memory.
Thus all label references were shifted by an amount equal to the number
of NOP instructions in the program.

Suppose, there is a JMP instruction to location 17 relative to the start
of the program. Since, NOP does not occupy any space the instruction
that is supposed to at 17 will be 16 the JMP will be invalid.